Refactor scheduler into separate file
